+++ src/components/Popups/TransactionPopup.tsx
@@ -1,29 +1,30 @@
 import { useWeb3React } from '@web3-react/core'
 import Column from 'components/Column'
-import { ActivityRowStyledDescriptor } from 'components/WalletDropdown/MiniPortfolio/Activity'
 import { parseLocalActivity } from 'components/WalletDropdown/MiniPortfolio/Activity/parseLocal'
 import { PortfolioLogo } from 'components/WalletDropdown/MiniPortfolio/PortfolioLogo'
 import PortfolioRow from 'components/WalletDropdown/MiniPortfolio/PortfolioRow'
 import useENSName from 'hooks/useENSName'
 import { useCombinedActiveList } from 'state/lists/hooks'
 import { useTransaction } from 'state/transactions/hooks'
-import { ThemedText } from 'theme'
+import { TransactionDetails } from 'state/transactions/types'
+import styled from 'styled-components/macro'
+import { EllipsisStyle, ThemedText } from 'theme'
 import { ExplorerDataType, getExplorerLink } from 'utils/getExplorerLink'
 
 import { PopupAlertTriangle } from './FailedNetworkSwitchPopup'
 
-export default function TransactionPopup({ hash }: { hash: string }) {
-  const { chainId } = useWeb3React()
+const Descriptor = styled(ThemedText.BodySmall)`
+  ${EllipsisStyle}
+`
 
-  const tx = useTransaction(hash)
-  const tokens = useCombinedActiveList()
+function TransactionPopupContent({ tx, chainId }: { tx: TransactionDetails; chainId: number }) {
   const success = Boolean(tx?.receipt && tx?.receipt.status === 1)
-
-  const activity = tx ? parseLocalActivity(tx, chainId ?? 0, tokens) : null
+  const tokens = useCombinedActiveList()
+  const activity = tx ? parseLocalActivity(tx, chainId, tokens) : null
   const { ENSName } = useENSName(activity?.otherAccount)
-  const explorerUrl = getExplorerLink(chainId ?? 0, hash, ExplorerDataType.TRANSACTION)
+  const explorerUrl = getExplorerLink(chainId, tx.hash, ExplorerDataType.TRANSACTION)
 
-  if (!chainId || !tx || !activity) return null
+  if (!activity) return null
 
   return (
     <PortfolioRow
@@ -43,13 +44,22 @@ export default function TransactionPopup({ hash }: { hash: string }) {
       }
       title={<ThemedText.SubHeader fontWeight={500}>{activity.title}</ThemedText.SubHeader>}
       descriptor={
-        <ActivityRowStyledDescriptor color="textSecondary">
+        <Descriptor color="textSecondary">
           {activity.descriptor}
           {ENSName ?? activity.otherAccount}
-        </ActivityRowStyledDescriptor>
+        </Descriptor>
       }
       onClick={() => window.open(explorerUrl, '_blank')}
-      right={undefined}
     />
   )
+}
+
+export default function TransactionPopup({ hash }: { hash: string }) {
+  const { chainId } = useWeb3React()
+
+  const tx = useTransaction(hash)
+
+  if (!chainId || !tx) return null
+
+  return <TransactionPopupContent tx={tx} chainId={chainId} />
 }
